refactor(cuts): extracted cuts from csg & moved to api
