+++ src/core/math/Side.js
@@ -55,7 +55,7 @@ Side.prototype = {
   },
 
   toPolygon3D: function (z0, z1) {
-    //console.log(this.vertex0.pos)
+    // console.log(this.vertex0.pos)
     const vertices = [
       new Vertex3(this.vertex0.pos.toVector3D(z0)),
       new Vertex3(this.vertex1.pos.toVector3D(z0)),
